Refactored LOTS
TONNES OF ERRORS
