+++ BitCalculator/Program.cs
@@ -27,7 +27,7 @@ namespace DevTools
                 _operator = @operator;
             }
         }
-        static ulong lastInput = 0ul;
+        public static ulong lastInput = 0ul;
         static bool defaultFlipVal = false;
         public static string lastprint;
         static void Main(string[] args)
@@ -79,97 +79,7 @@ namespace DevTools
         }
 
         public static bool expectingError; //Set this to true, and it means the application is throwing a custom error
-        //This will not print out a stack trace
-
-        public static bool readingConsole = false; //Staticbool that shows for async methods whether we are currently reading text from the console
-        static string retString = ""; //Buffer for the readline function
-        private static string ReadLineOrEsc()
-        {
-            retString = "";
-            readingConsole = true;
-            int writeIDX = 0;
-
-            do
-            {
-                ConsoleKeyInfo readKeyResult = Console.ReadKey(true);
-
-                if (readKeyResult.Key == ConsoleKey.Enter) //Return the buffer
-                {
-                    Console.WriteLine();
-                    readingConsole = false;
-                    return retString;
-                }
-
-                if (readKeyResult.Key == ConsoleKey.LeftArrow) //Move left
-                {
-                    writeIDX--;
-                    Console.CursorLeft--;
-                }
-                else if (readKeyResult.Key == ConsoleKey.RightArrow) //Mofe right
-                {
-                    if (retString.LettersLength()>writeIDX)//Not at end yet?
-                    {
-                        writeIDX++;
-                        Console.CursorLeft++;
-                    }
-                }
-                else if (readKeyResult.Key == ConsoleKey.UpArrow || readKeyResult.Key == ConsoleKey.DownArrow || readKeyResult.Key == ConsoleKey.Delete || readKeyResult.Key == ConsoleKey.Tab)
-                {
-                    //Do nothing, just dont run other functions
-
-                    //Delete does not work in this application as it moves the cursor uncrontrollably
-                    //To avoid this error, we just don't cater for it
-                }
-
-                // handle backspace
-                else if (readKeyResult.Key == ConsoleKey.Backspace)
-                {
-                    if (writeIDX != 0) //Don't delete characters before the first one
-                    {
-                        retString = retString.Remove(writeIDX - 1, 1); //Remove current char from the buffer
-                        ChangeUserTextColourLive(retString); //Change user text colour
-                        Console.CursorLeft--; //Move the cursor to its new position
-                        writeIDX--; //Update current writeIDX
-                    }
-                }
-                else
-                {
-                    if (retString.Length == writeIDX)//Writing next character?
-                    {
-                        retString += readKeyResult.KeyChar; //Add to the buffer
-                        Console.CursorLeft++; //Move the cursor right
-                        writeIDX++;
-                    }
-                    else if (writeIDX >= 0)//We have moved the idx?
-                    {
-                        StringBuilder sb = new StringBuilder(retString);
-                        sb[writeIDX] = readKeyResult.KeyChar; //Replace specific position in string
-                        retString = sb.ToString(); //Update string
-
-                        writeIDX++;
-                        Console.CursorLeft++;
-                    }
-                    ChangeUserTextColourLive(retString); //Change the colour of the users text
-                }
-                if (Console.CursorLeft <= 3)
-                {
-                   Console.CursorLeft = 3;
-                    writeIDX = 0;
-                }
-            }
-            while (true);
-        }
-
-        private static void SetupConsole()
-        {
-            Colorful.Console.BackgroundColor = Color.FromArgb(0, 16, 29); //Change the background colour to the snazzy blue
-
-            Colorful.Console.Clear();
-            Colorful.Console.OutputEncoding = Encoding.Unicode;
-            Colorful.Console.WriteAsciiStyled("Dev Tools 2022", new Colorful.StyleSheet(Color.FromArgb(122, 224, 255)));
-            Colorful.Console.WriteLine("Type help to show all functions", Color.FromArgb(122, 224, 255));
-            Colorful.Console.ForegroundColor = Color.FromArgb(10, 181, 158);
-        }
+        //This will not print out a stack traces
 
         /// <summary>
         /// Checks to see if directories are valid. re-creates files if nessecary
@@ -212,54 +122,6 @@ namespace DevTools
             }
         }
 
-        private static string RemoveLineBreaks(string v)
-        {
-            return Regex.Replace(Regex.Replace(v, "\n", ""), "\r", "");
-        }
-
-        private static string RemoveSpaces(string input)
-        {
-            string result = "";
-            bool inSpeech = false;
-            bool isComment = false;
-            string buffer = "";
-            for (int i = 0; i < input.Length; ++i)
-            {
-                if (buffer.Contains(@"///"))
-                {
-                    result += input.Substring(i);
-                    break;
-                }
-                if (buffer.Contains("//"))
-                {
-                    isComment = true;
-                }
-                if (buffer.Contains("\\\\"))
-                {
-                    isComment = false;
-                    buffer = "";
-                }
-                char c = input[i];
-                if (c == '\"')
-                {
-                    inSpeech = !inSpeech;
-                    continue;
-                }
-                if (!inSpeech && !isComment)
-                {
-                    if (c != ' ')
-                    {
-                        result += c.ToString().ToLower()[0];
-                    }
-                }
-                else
-                {
-                    result += c;
-                }
-                buffer += c;
-            }
-            return result;
-        }
         public static void DoMainMethod(string userINPUT, bool removeSpaces = true)
         {
             if (userINPUT.BeginsWith("help-")) //Show help for specific function. Specified after the -
@@ -302,88 +164,6 @@ namespace DevTools
         /// Write snazzy text in the console that is big
         /// </summary>
         /// <param name="text">text to print</param>
-        public static void WriteAscii(string text)
-        {
-            Colorful.Console.WriteAsciiStyled(text, new Colorful.StyleSheet(Color.FromArgb(122, 224, 255)));
-        }
-        public static string FactoriseCrissCross(int first, int second, int c, bool write = true)
-        {
-            foreach (var factor in GetFactors(first))
-            {
-                int topleft = factor;
-                int bottomleft = first / factor;
-                foreach (var factor2 in GetFactors(c))
-                {
-                    int topright = factor2;
-                    int bottomright = c / factor2;
-                    if (bottomleft * topright + topleft * bottomright == second)
-                    {
-                        string co1 = topleft == 1 ? "" : topleft.ToString();
-                        string co2 = bottomleft == 1 || bottomleft == -1 ? "" : bottomleft.ToString();
-
-                        co1 = topleft == -1 ? "-" : co1;
-                        co2 = bottomleft == -1 ? "-" : co2;
-                        if (write == true)
-                        {
-                            string s_topleft = topleft.ToString();
-                            string s_topright = topright.ToString();
-                            s_topright = s_topright[0] == '-' ? s_topright : s_topright.Insert(0,"+");
-
-                            string s_bottomleft = bottomleft.ToString();
-                            string s_bottomright = bottomright.ToString();
-                            s_bottomright = s_bottomright[0] == '-' ? s_bottomright : s_bottomright.Insert(0, "+");
-
-
-                            PrintColour(string.Format("({0}x{1})({2}x{3})", s_topleft,s_topright, s_bottomleft, s_bottomright));
-                        }
-                        /**
-                         * 3    5
-                         *   --
-                         * 4    5
-                         * */
-                        return string.Format("({0}x {4} {1})({2}x {5} {3})", co1, Math.Abs(topright), co2, Math.Abs(bottomright), topright >= 0 ? "+" : "-", bottomright >= 0 ? "+" : "-");
-                    }
-                }
-            }
-            expectingError = true;
-            throw new Exception("Could not factorize");
-        }
-        static IEnumerable<int> GetFactors(int n)
-        {
-            for (int i = 1; i <= Math.Abs(n); ++i)
-            {
-                if (n % i == 0)
-                {
-                    yield return i;
-                    yield return -i;
-                }
-            }
-        }
-        public static bool VariableExists(string name)
-        {
-            foreach (var line in DefineVariableContents())
-            {
-                if (line.Split(',')[0] == name) //Does it exist?
-                {
-                    return true;
-                }
-            }
-            foreach (var line in File.ReadAllLines(FuncFilePath))
-            {
-                if (line.Split('(')[0] == name) //Does it exist?
-                {
-                    return true;
-                }
-            }
-            foreach (var variable in tempVariables.Keys)
-            {
-                if (variable == name) //Does it exist?
-                {
-                    return true;
-                }
-            }
-            return false;
-        }
         static bool noprint = false;
         public static void MainMethod(string userINPUT, bool removeSpaces = true)
         {
@@ -779,72 +559,6 @@ namespace DevTools
             }
         }
 
-        private static void PrintIPData()
-        {
-            Console.Write("Local IP: ");
-            NetworkingPrint(GetLocalIPAddress());
-
-            NetworkInterface[] interfaces = NetworkInterface.GetAllNetworkInterfaces();
-            foreach (NetworkInterface adapter in interfaces)
-            {
-                Console.Write("Name: ");
-                NetworkingPrint(adapter.Name);
-                Console.WriteLine(adapter.Description);
-                NetworkingPrint(String.Empty.PadLeft(adapter.Description.Length, '='));
-                NetworkingPrint("  Interface type .......................... : "+ adapter.NetworkInterfaceType);
-                NetworkingPrint("  Operational status ...................... : "+adapter.OperationalStatus);
-                string versions = "";
-
-                // Create a display string for the supported IP versions.
-                if (adapter.Supports(NetworkInterfaceComponent.IPv4))
-                {
-                    versions = "IPv4";
-                }
-                if (adapter.Supports(NetworkInterfaceComponent.IPv6))
-                {
-                    if (versions.Length > 0)
-                    {
-                        versions += " ";
-                    }
-                    versions += "IPv6";
-                }
-                NetworkingPrint("  IP version .............................. : " +versions);
-                Console.WriteLine();
-            }
-        }
-        public static bool NetworkingPrint(string toprint)
-        {
-            if (readingConsole) //Are we currently reading user input?
-            {
-                var left = Console.CursorLeft;
-                Console.CursorLeft = 0;
-                ClearCurrentConsoleLine();
-                Colorful.Console.WriteLine(toprint, Color.FromArgb(184, 186, 255));
-                Colorful.Console.Write("-->", Color.FromArgb(10, 181, 158)); //Header for text
-
-                PrintColour(retString, false, false, false);
-                Console.CursorLeft = left; //Reset the left to its original position
-            }
-            else
-            {
-                Colorful.Console.WriteLine(toprint, Color.FromArgb(184, 186, 255));
-            }
-            return false;
-        }
-        public static string GetLocalIPAddress()
-        {
-            var host = Dns.GetHostEntry(Dns.GetHostName());
-            foreach (var ip in host.AddressList)
-            {
-                if (ip.AddressFamily == AddressFamily.InterNetwork)
-                {
-                    return ip.ToString();
-                }
-            }
-            expectingError = true;
-            throw new Exception("No network adapters with an IPv4 address in the system!");
-        }
-
         /// <summary>
         /// Weird math thingy
         /// </summary>
@@ -861,7 +575,7 @@ namespace DevTools
         /// CD = 12
         /// DB = 8
         /// etc....
-        /// I have no idea why I implemented this.....
+        /// This is not designed to be a USEFUL tool, more just something to play around with
         static char[] alphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZ".ToCharArray(); //Static string of all the letters in the alphabet
         private static void PrintAti(string userINPUT)
         {
@@ -883,31 +597,6 @@ namespace DevTools
             }
             Colorful.Console.WriteAscii(string.Format("{0:n0}", total)); //Print result as snazzy asci text
         }
-        /// <summary>
-        /// Function searches for // to lead a comment and \\ to exit one
-        /// It prints these to the console and removes them from the input string
-        /// 
-        /// It searches for the first versions of // and \\
-        /// Removes them from the string, then recurses to see if there is a second set
-        /// </summary>
-        /// <param name="sINPUT"></param>
-        /// <returns></returns>
-        private static string RemoveComments(string sINPUT)
-        {
-            if (sINPUT.Contains(@"//") && sINPUT.Contains(@"\\") && !sINPUT.Contains("#defunc"))
-            {
-                string comment = sINPUT.Substring(sINPUT.IndexOf(@"//") + 2, sINPUT.IndexOf(@"\\") - sINPUT.IndexOf(@"//") - 2);
-                //comment is the substring, +2 is to removing the leading //
-                //-2 is to remove the following \\
-                Colorful.Console.WriteLine(comment, Color.Beige); //Write the comment to the console
-                return RemoveComments(sINPUT.Substring(0, sINPUT.IndexOf(@"//")) + sINPUT.Substring(sINPUT.IndexOf(@"\\") + 2));
-                //use recursion to see if there are any more comments
-            }
-            else
-            {
-                return sINPUT;
-            }
-        }
         public static string RemoveX(string userINPUT)
         {
             userINPUT = ReplaceTempVariables(userINPUT);
@@ -919,123 +608,6 @@ namespace DevTools
 
             userINPUT = RemoveBooleanStatements(userINPUT);
             return userINPUT;
-        }
-        public static string RemoveLog(string userINPUT)
-        {
-            List<int> logidxs = userINPUT.AllIndexs("log"); //Find the positions of all the log statements
-            var idx = userINPUT.IndexOf("log");
-            if (idx == -1)
-            {
-                return userINPUT;
-            }
-            int openingBracketIDX = NextBracket(userINPUT, idx);
-
-            var logbase = userINPUT.Substring(idx + 3, openingBracketIDX - idx - 3); //+3 is for the length of log
-            if (logbase == "")
-            {
-                logbase = "10"; //Default base is 10
-            }
-
-            var lognum = userINPUT.Substring(openingBracketIDX + 1, ClosingBracket(userINPUT, openingBracketIDX + 1) - openingBracketIDX - 1);
-
-            string before = userINPUT.Substring(0, idx); //Get the string that comes before this, up until idx
-            string after = userINPUT.Substring(ClosingBracket(userINPUT, openingBracketIDX + 1)+1); //End at the closing bracket
-            string replace = Math.Log(double.Parse(lognum), double.Parse(logbase)).ToString();
-            userINPUT = before + replace + after; //Modify the string
-
-            userINPUT = RemoveLog(userINPUT);
-
-            if (userINPUT.StartsWith("np")) //User doesn't want to print binary of the result
-            {
-                userINPUT = userINPUT.Substring(2);
-                noprint = true;
-            }
-            if (!userINPUT.StartsWith("doum"))
-            {
-                userINPUT = userINPUT.Insert(0, "doum");
-            }
-
-            return userINPUT;
-        }
-        /// <summary>
-        /// Descriptions are marked with ///
-        /// This function prints the descriptions to the console
-        /// 
-        /// Mark with ** where you want things to be recognised as normal commands
-        /// </summary>
-        /// <param name="sINPUT"></param>
-        /// <returns></returns>
-        enum PrintType
-        {
-            Comment,
-            Command,
-        }
-        private static string ShowDescription(string sINPUT)
-        {
-            sINPUT = Regex.Replace(sINPUT, @"<new>", "\n"); //Replace all \n with new lines
-
-            if (sINPUT.Contains(@"///"))
-            {
-                string comment = sINPUT.Substring(sINPUT.IndexOf(@"///") + 3); //Find the end of the first ///. +3 to get to the end of the "///"
-                List<Comment> toprint = new List<Comment>();
-                string buffer = "";
-                bool lookingForAsterix = false;
-                for (int i = 0; i < comment.Length; i++)
-                {
-                    char c = comment[i];
-                    if (c == '\\') //Is it a \
-                                   //Uses \\, but is really looking for \
-                    {
-                        toprint.Add(new Comment(buffer, PrintType.Comment)); //Add this as being normal
-                        buffer = "";
-                        break; //UNTESTED IMPROVEMENT
-                    }
-                    else if (c == '*') //Asterix shows we want to print out as code
-                    {
-                        if (lookingForAsterix) //Are we looking for the ending asterix?
-                        {
-                            toprint.Add(new Comment(buffer, PrintType.Command)); //Add this as being code
-                            lookingForAsterix = false;
-                            buffer = "";
-                        }
-                        else
-                        {
-                            toprint.Add(new Comment(buffer, PrintType.Comment)); //Add this as being normal
-                            lookingForAsterix = true;
-                            buffer = "";
-                        }
-                    }
-                    else
-                    {
-                        buffer += c;
-                    }
-                }
-                foreach (var c in toprint)
-                {
-                    switch (c.printType)
-                    {
-                        case PrintType.Comment:
-                            Colorful.Console.Write(c.s, Color.Beige);
-                            break;
-                        case PrintType.Command:
-                            PrintColour(c.s, false, false, false);
-                            break;
-                    }
-                }
-                Console.WriteLine();
-            }
-            return sINPUT;
-        }
-        struct Comment
-        {
-            public string s;
-            public PrintType printType;
-
-            public Comment(string s, PrintType printType)
-            {
-                this.s = s;
-                this.printType = printType;
-            }
         }
         /// <summary>
         /// Runs a custom system function ran()
@@ -1078,82 +650,6 @@ namespace DevTools
             }
             return sINPUT;
         }
-
-        #region printDecimals
-        /// <summary>
-        /// Prints a double decimal values binary output
-        /// So far NO bugs exist. Will not be documented, because I have no f*cking idea how it works
-        /// </summary>
-        /// <param name="input"></param>
-        private static void PrintDouble(string input)
-        {
-            var res = Regex.Replace(RemoveSpaces(input), "\n", ""); //Remove all spaces, new lines or blanks from string
-
-
-            for (int i = 0; i < 64; ++i)
-            {
-                if (i == 0)
-                {
-                    Colorful.Console.Write(res[i], Color.FromArgb(255, 255, 255));
-                }
-                else if (i <= 11)
-                {
-                    Colorful.Console.Write(res[i], Color.FromArgb(0, 255, 10));
-                }
-                else
-                {
-                    Colorful.Console.Write(res[i], Color.FromArgb(255, 100, 100));
-                }
-                if ((i + 1) % 8 == 0 && i != 0)
-                {
-                    Colorful.Console.Write('\n');
-                }
-                else
-                {
-                    Colorful.Console.Write(' ');
-                }
-            }
-            Colorful.Console.WriteLine("This colour is the mantissa", Color.FromArgb(255, 100, 100));
-            Colorful.Console.WriteLine("This colour is the exponent", Color.FromArgb(0, 255, 10));
-            Colorful.Console.WriteLine("This colour is the sign", Color.FromArgb(255, 255, 255));
-        }
-        /// <summary>
-        /// Prints a float decimal values binary output
-        /// So far NO bugs exist. Will not be documented, because I have no f*cking idea how it works
-        /// </summary>
-        /// <param name="input"></param>
-        private static void PrintFloat(string input)
-        {
-            var res = Regex.Replace(RemoveSpaces(input), "\n", "");
-            for (int i = 0; i < 32; ++i)
-            {
-                if (i == 0)
-                {
-                    Colorful.Console.Write(res[i], Color.FromArgb(255, 255, 255));
-                }
-                else if (i <= 9)
-                {
-                    Colorful.Console.Write(res[i], Color.FromArgb(0, 255, 10));
-                }
-                else
-                {
-                    Colorful.Console.Write(res[i], Color.FromArgb(255, 100, 100));
-                }
-                if ((i + 1) % 8 == 0 && i != 0)
-                {
-                    Colorful.Console.Write('\n');
-                }
-                else
-                {
-                    Colorful.Console.Write(' ');
-                }
-            }
-            //Colorful.Console.BackgroundColor = Color.FromArgb(0, 16, 29);
-            Colorful.Console.WriteLine("This colour is the mantissa", Color.FromArgb(255, 100, 100));
-            Colorful.Console.WriteLine("This colour is the exponent", Color.FromArgb(0, 255, 10));
-            Colorful.Console.WriteLine("This colour is the sign", Color.FromArgb(255, 255, 255));
-        }
-        #endregion
 
 
         /// <summary>
@@ -1271,183 +767,6 @@ namespace DevTools
             return sINPUT;
         }
 
-        #region angleconversions
-        public static double DegreeToRadian(double angle)
-        {
-            return angle * (Math.PI / 180.0);
-        }
-        public static double RadianToDegree(double angle)
-        {
-            return angle * (180.0 / Math.PI);
-        }
-        #endregion
-
-        /// <summary>
-        /// Removes trig statesments from the users input
-        /// 
-        /// IS BUGGY, REQUIRES FIX
-        /// </summary>
-        /// <param name="sINPUT"></param>
-        /// <returns></returns>
-        private static string RemoveTrig(string sINPUT)
-        {
-            if (sINPUT.Contains("sin(") ||
-                sINPUT.Contains("arcsin(") ||
-                sINPUT.Contains("tan(") ||
-                sINPUT.Contains("arctan(") ||
-                sINPUT.Contains("cos(") ||
-                sINPUT.Contains("arccos("))
-            {
-                string buffer = "";
-                for (int i = 0; i < sINPUT.Length; i++)
-                {
-                    char c = sINPUT[i];
-                    buffer += c;
-                    if (buffer.Contains("arcsin("))
-                    {
-                        return CalcTrig(sINPUT, i, MathAngleType.ArcSin);
-                    }
-                    if (buffer.Contains("arccos("))
-                    {
-                        return CalcTrig(sINPUT, i, MathAngleType.ArcCos);
-                    }
-                    if (buffer.Contains("arctan("))
-                    {
-                        return CalcTrig(sINPUT, i, MathAngleType.ArcTan);
-                    }
-                    if (buffer.Contains("sin("))
-                    {
-                        return CalcTrig(sINPUT, i, MathAngleType.Sin);
-                    }
-                    if (buffer.Contains("cos("))
-                    {
-                        return CalcTrig(sINPUT, i, MathAngleType.Cos);
-                    }
-                    if (buffer.Contains("tan("))
-                    {
-                        return CalcTrig(sINPUT, i, MathAngleType.Tan);
-                    }
-                }
-            }
-            return sINPUT;
-        }
-        enum MathAngleType
-        {
-            Sin,
-            Cos,
-            Tan,
-            ArcSin,
-            ArcCos,
-            ArcTan,
-        }
-        private static string CalcTrig(string sINPUT, int stringIDX, MathAngleType mathAngleType)
-        {
-            string fixedval = sINPUT.Substring(0, stringIDX - 3); //Find the math that happens before it
-            int nextOperaror = ClosingBracket(sINPUT, stringIDX + 1); //The next operator
-            //BUG IF IT INCLUDES E-...d
-            string result = DoubleRemoveBrackets(sINPUT.Substring(stringIDX + 1, nextOperaror - stringIDX - 1)); //Find the number between the brackets
-            double calcNum = DegreeToRadian(double.Parse(result)); //Convert it from degrees to radians
-            switch (mathAngleType)
-            {
-                case MathAngleType.Sin:
-                    calcNum = Math.Sin(calcNum);
-                    break;
-                case MathAngleType.Cos:
-                    calcNum = Math.Cos(calcNum);
-                    break;
-                case MathAngleType.Tan:
-                    calcNum = Math.Tan(calcNum);
-                    break;
-                case MathAngleType.ArcSin:
-                    calcNum = Math.Asin(calcNum);
-                    break;
-                case MathAngleType.ArcCos:
-                    calcNum = Math.Acos(calcNum);
-                    break;
-                case MathAngleType.ArcTan:
-                    calcNum = Math.Atan(calcNum);
-                    break;
-            }
-            string afterThat = sINPUT.Substring(nextOperaror + 1, sINPUT.Length - nextOperaror - 1);
-            PrintColour(string.Format("{0}({1}) = {2}",mathAngleType.ToString() , result, calcNum), true);
-            string return_result = fixedval + calcNum + afterThat;
-
-            if (return_result.StartsWith("arc")) //Was this an arc operation?
-            {
-                return_result = return_result.Substring(3); //Remove it
-            }
-            if (return_result.StartsWith("nw"))
-            {
-                printWorkings = false;
-                return_result = return_result.Substring(2);
-            }
-            if (return_result.StartsWith("np"))
-            {
-                noprint = true;
-                return_result = return_result.Substring(2);
-            }
-            if (return_result.Length <= 3 || return_result.Substring(0,4) != "doum")
-            {
-                return_result = return_result.Insert(0,"doum");
-            }
-            return RemoveTrig(return_result);
-        }
-
-        /// <summary>
-        /// Finds the index of the closing bracket in a string
-        /// NO BUGS HERE
-        /// </summary>
-        /// <param name="sINPUT"></param>
-        /// <param name="curridx">where to search from in the string</param>
-        /// <returns></returns>
-        private static int ClosingBracket(string sINPUT, int curridx)
-        {
-            int amountOfOpenBrackets = 1; //Assume that whatever is giving this string to us has alredy found an open bracket
-            int amountOfClosingBrackets = 0; //We will need to look for the closing bracket
-            for (int i = curridx; i < sINPUT.Length; ++i) //Iterate through the string from the curridx
-            {
-                if (sINPUT[i] == '(') //Is there a set of nested brackets?
-                {
-                    amountOfOpenBrackets++; //Increase the amount of open brackets, so that we have to find MORE closing brackets
-                }
-                if (sINPUT[i] == ')')
-                {
-                    amountOfClosingBrackets++; //One set of brackets has been closed
-                }
-                if (amountOfClosingBrackets == amountOfOpenBrackets) //Is there the same amount of open brackets as there are closing brackets
-                                                                     //This symbolizes that all nested sets of brackets have been found
-                {
-                    return i; //Return the idx of the last bracket
-                }
-            }
-            return sINPUT.Length - 1; //No closing bracket was found! just return the end of the string
-        }
-        /// <summary>
-        /// Deletes the variable as the name specifies
-        /// </summary>
-        /// <param name="input"></param>
-        private static void DeleteVariable(string input)
-        {
-            List<string> variables = File.ReadAllLines(DataFilePath).ToList(); //Find a list of all the variables
-            var copy = new List<string>();
-            foreach (var v in variables)
-            {
-                copy.Add(v);
-            } //Copy the list so that we can iterate through this list while removing items from the other
-            foreach (var s in variables)
-            {
-                var v = s.Split(',')[0];
-                if (v == input) //Is this the variable we are deleting?
-                {
-                    copy.Remove(s); //Remove it from the list
-                }
-            }
-            foreach (var v in copy)
-            {
-                PrintColour(v, true); //Print out the rest of the variables defined
-            }
-            File.WriteAllLines(DataFilePath, copy); //Write the new variable set to the file
-        }
         /// <summary>
         /// Replaces binary as defined by b_010101  with its corresponding int value
         /// </summary>
@@ -1515,14 +834,6 @@ namespace DevTools
             return input;
         }
         /// <summary>
-        /// Prints Hex in a snazzy colour
-        /// </summary>
-        /// <param name="v"></param>
-        private static void PrintHex(string v)
-        {
-            Colorful.Console.WriteLine(v.Insert(0, "#"), Color.FromArgb(234, 255, 0));
-        }
-        /// <summary>
         /// Converts hex to rgb
         /// You type in hrgb #ffffff
         /// An voila, it converts it into rgb(255,255,255)
@@ -1553,467 +864,6 @@ namespace DevTools
             result += ");";
             PrintColour(result, false);
         }
-
-        struct FuncLocation
-        {
-            public int start;
-            public int end;
-            public string name;
-
-            public FuncLocation(int start, int end, string name)
-            {
-                this.start = start;
-                this.end = end;
-                this.name = name;
-            }
-        }
-
-        /// <summary>
-        /// An extension method for Colourful.Console.Writeline()
-        /// </summary>
-        /// <param name="v"></param>
-        /// <param name="workings"></param>
-        /// <param name="isBin"></param>
-        public static void PrintColour(string v, bool workings = false, bool isBin = false, bool writeline = true)
-        {
-            //Checking for functions
-            #region functionCheck
-            var AllFunctions = File.ReadAllLines(FuncFilePath);
-            List<FuncLocation> funclocations = new List<FuncLocation>();
-            foreach (var s in File.ReadAllLines(FuncFilePath))
-            {
-                if (s == "")
-                {
-                    continue;
-                }
-                if (s == "SYSTEM FUNCTIONS:")
-                {
-                    break;
-                }
-                var name = s.Split('(')[0];
-                if (v.Contains(name) && name.Length >= 2) //Name is in the users input?
-                {
-                    foreach (var idx in v.AllIndexs(name))
-                    {
-                        if (idx == 0 || !char.IsLetter(v[idx - 1])) //Is this the start of the word?
-                        {
-                            if (idx + name.Length < v.Length && !char.IsLetter(v[idx + name.Length]))//Is this the end of a word?
-                            {
-                                funclocations.Add(new FuncLocation(v.IndexOf(name), ClosingBracket(v, v.IndexOf(name) + name.Length + 1), name));
-                            }
-                        }
-                    }
-                }
-            }
-
-            bool atsystem = false;
-            List<FuncLocation> sysfunclocations = new List<FuncLocation>();
-            foreach (var s in File.ReadAllLines(FuncFilePath))
-            {
-                if (s == "SYSTEM FUNCTIONS:")
-                {
-                    atsystem = true;
-                    continue;
-                }
-                if (!atsystem)
-                {
-                    continue;
-                }
-                var name = s.Split('(')[0];
-                if (v.Contains(name) && name.Length >= 1) //Name is in the users input?
-                {
-                    foreach (var idx in v.AllIndexs(name))
-                    {
-                        if (idx == 0 || !char.IsLetter(v[idx-1])) //Is this the start of the word?
-                        {
-                            if (idx + name.Length >= v.Length || !char.IsLetter(v[idx + name.Length]))//Is this the end of a word?
-                            {
-                                sysfunclocations.Add(new FuncLocation(idx, idx + name.Length, name));
-                            }
-                        }
-                    }
-                }
-            }
-
-            List<FuncLocation> variableLocations = new List<FuncLocation>();
-            foreach (var s in File.ReadAllLines(DataFilePath))
-            {
-                if (!s.Contains(',')) //No comma in the line?
-                {
-                    File.WriteAllText(DataFilePath, ""); //Clear the file
-                    expectingError = true;
-                    throw new Exception("Variables file corrupted. File cleared");
-                }
-                var name = s.Split(',')[0];
-                if (v.Contains(name)) //Name is in the users input?
-                {
-                    foreach (var idx in v.AllIndexs(name))
-                    {
-                        if (idx == 0 || !char.IsLetter(v[idx - 1])) //Is this the start of the word?
-                        {
-                            if (idx + name.Length >= v.Length || !char.IsLetter(v[idx + name.Length]))//Is this the end of a word?
-                            {
-                                variableLocations.Add(new FuncLocation(idx, idx + name.Length, name));
-                            }
-                        }
-                    }
-                }
-            }
-            foreach (var name in tempVariables.Keys)
-            {
-                if (v.Contains(name)) //Name is in the users input?
-                {
-                    foreach (var idx in v.AllIndexs(name))
-                    {
-                        if (idx == 0 || !char.IsLetter(v[idx - 1])) //Is this the start of the word?
-                        {
-                            if (idx + name.Length >= v.Length || !char.IsLetter(v[idx + name.Length]))//Is this the end of a word?
-                            {
-                                variableLocations.Add(new FuncLocation(idx, idx + name.Length, name));
-                            }
-                        }
-                    }
-                }
-            }
-            foreach (var name in networkingVariables.Keys)
-            {
-                if (v.Contains(name)) //Name is in the users input?
-                {
-                    foreach (var idx in v.AllIndexs(name))
-                    {
-                        if (idx == 0 || !char.IsLetter(v[idx - 1])) //Is this the start of the word?
-                        {
-                            if (idx + name.Length >= v.Length || !char.IsLetter(v[idx + name.Length]))//Is this the end of a word?
-                            {
-                                variableLocations.Add(new FuncLocation(idx, idx + name.Length, name));
-                            }
-                        }
-                    }
-                }
-            }
-
-            List<FuncLocation> speechLocations = new List<FuncLocation>();
-            bool lookingForSpeech = false;
-            int lastidx = -1;
-            for (int i = 0; i < v.Length; ++i)
-            {
-                char c = v[i];
-                if (c == '\"') //In speechmarks?
-                {
-                    if (lookingForSpeech)
-                    {
-                        lookingForSpeech = false;
-                        speechLocations.Add(new FuncLocation(lastidx, i+1, ""));
-                        lastidx = -1;
-                    }
-                    else
-                    {
-                        lastidx = i;
-                        lookingForSpeech = true;
-                    }
-                }
-            }
-            if (lastidx != -1) //Is there a lonely speech mark?
-            {
-                speechLocations.Add(new FuncLocation(lastidx, v.Length, "")); //All text after that speech mark is marked as text   
-            }
-            #endregion
-            funclocations = funclocations.OrderBy(f => f.start).ToList();
-            sysfunclocations = sysfunclocations.OrderBy(f => f.start).ToList();
-            variableLocations = variableLocations.OrderBy(f => f.start).ToList();
-            speechLocations = speechLocations.OrderBy(f => f.start).ToList();
-            //Order the functions so that the ones that appear first are at the start of the list
-
-            if (workings && printWorkings == false)
-            {
-                return;
-            }
-            if (isBin)
-            {
-                foreach (var c in v)
-                {
-                    if (c == '0')
-                    {
-                        Colorful.Console.Write(c, Color.FromArgb(90, 255, 183));
-                    }
-                    else
-                    {
-                        Colorful.Console.Write(c, Color.FromArgb(10, 181, 158));
-                    }
-                } //Binary font
-            }
-            else
-            {
-                bool isyellow = false;
-                bool printingFunction = false;
-                bool printingsysFunction = false;
-                bool printingDefine = false;
-                bool printingSpeech = false;
-                FuncLocation current = new FuncLocation(int.MaxValue, int.MaxValue, "");
-                FuncLocation currentsys = new FuncLocation(int.MaxValue, int.MaxValue, "");
-                FuncLocation currentdefine = new FuncLocation(int.MaxValue, int.MaxValue, "");
-                FuncLocation currentspeech = new FuncLocation(int.MaxValue, int.MaxValue, "");
-                for (int i = 0; i < v.Length; i++)
-                {
-                    if (currentsys.end == i)
-                    {
-                        printingsysFunction = false;
-                    }
-                    if (currentdefine.end == i)
-                    {
-                        printingDefine = false;
-                    }
-                    if (currentspeech.end == i)
-                    {
-                        printingSpeech = false;
-                    }
-
-                    if (funclocations.Count != 0 && funclocations[0].start == i) //Are we at the start idx of a function?
-                    {
-                        current = funclocations[0];
-                        printingFunction = true;
-                        funclocations.RemoveAt(0); //Remove it from the list
-                    }
-                    if (sysfunclocations.Count != 0 && sysfunclocations[0].start == i) //Are we at the start idx of a system function?
-                    {
-                        currentsys = sysfunclocations[0];
-                        printingsysFunction = true;
-                        sysfunclocations.RemoveAt(0); //Remove it from the list
-                    }
-                    if (variableLocations.Count != 0 && variableLocations[0].start == i) //Are we at the start idx of a system function?
-                    {
-                        currentdefine = variableLocations[0];
-                        printingDefine = true;
-                        variableLocations.RemoveAt(0); //Remove it from the list
-                    }
-                    if (speechLocations.Count != 0 && speechLocations[0].start == i) //Are we at the start idx of a system function?
-                    {
-                        currentspeech = speechLocations[0];
-                        printingSpeech = true;
-                        speechLocations.RemoveAt(0); //Remove it from the list
-                    }
-
-                    char c = v[i];
-
-                    if (printingFunction) //Are we printing out a function?
-                    {
-                        if (i < current.start + current.name.Length) //Are we in the name of the function?
-                        {
-                            Colorful.Console.Write(c, Color.FromArgb(168, 0, 149));
-                        }
-                        else if (i == current.start + current.name.Length && c== '(') //Opening bracket?
-                        {
-                            Colorful.Console.Write(c, Color.FromArgb(255,255,255));
-                        }
-                        else if (i == current.end && c == ')') //Closing bracket?
-                        {
-                            Colorful.Console.Write(c, Color.FromArgb(255, 255, 255));
-                        }
-                        else
-                        {
-                            Colorful.Console.Write(c, Color.FromArgb(10, 181, 158));
-                        }
-                    }
-                    else if (printingsysFunction)
-                    {
-                        Colorful.Console.Write(c, Color.FromArgb(247, 255, 161));
-                    }
-                    else if (printingSpeech)
-                    {
-                        Colorful.Console.Write(c, Color.Beige);
-                    }
-                    else if (printingDefine)
-                    {
-                        Colorful.Console.Write(c, Color.FromArgb(168, 0, 149));
-                    }
-                    else
-                    {
-                        if (isyellow && !IsOperator(c) && c != ' ')
-                        {
-                            Colorful.Console.Write(c, Color.FromArgb(234, 255, 0));
-                            continue;
-                        }
-
-                        if (char.IsNumber(c))
-                        {
-                            Colorful.Console.Write(c, Color.FromArgb(6, 153, 255));
-                        }
-                        else if (c == '#')
-                        {
-                            isyellow = true;
-                            Colorful.Console.Write(c, Color.FromArgb(234, 255, 0));
-                        }
-                        else if (IsOperator(c) || c == ' ')
-                        {
-                            isyellow = false;
-                            Colorful.Console.Write(c, Color.FromArgb(130, 253, 255));
-                        }
-                        else
-                        {
-                            Colorful.Console.Write(c, Color.FromArgb(10, 181, 158));
-                        }
-                    }
-                    if (current.end == i) //Are we at the end of a function?
-                    {
-                        printingFunction = false;
-                    }
-                }
-            }
-            if (writeline)
-            {
-                Colorful.Console.WriteLine();
-            }
-            lastprint = v;
-        }
-        public static void ChangeUserTextColour(string userinput)
-        {
-            var x = Console.CursorLeft;
-            var y = Console.CursorTop;
-            y--; //Get the row above us
-            Console.SetCursorPosition(x,y);
-            Console.WriteLine(); //Clear previous line
-            Console.SetCursorPosition(x, y);
-
-            Colorful.Console.Write("-->", Color.FromArgb(255, 181, 158)); //Re-print the header
-
-            PrintColour(userinput);
-        }
-        public static void ClearCurrentConsoleLine()
-        {
-            int currentLineCursor = Console.CursorTop;
-            Console.SetCursorPosition(0, Console.CursorTop);
-            Console.Write(new string(' ', Console.WindowWidth));
-            Console.SetCursorPosition(0, currentLineCursor);
-        }
-        public static void ChangeUserTextColourLive(string userinput)
-        {
-            var x = Console.CursorLeft;
-            var y = Console.CursorTop;
-
-            Console.SetCursorPosition(0, y);
-            ClearCurrentConsoleLine();
-            Colorful.Console.Write("-->", Color.FromArgb(10, 181, 158)); //Header for text
-
-            Console.SetCursorPosition(3, y);
-
-            PrintColour(userinput, false, false, false);
-            Console.SetCursorPosition(x, y);
-        }
-        public static bool printWorkings = true;
-        public static void PrintError(string errorMessage)
-        {
-            Colorful.Console.WriteLine(errorMessage, Color.Red);
-        }
-        public static bool IsOperator(char c)
-        {
-            return c switch
-            {
-                '+' => true,
-                '-' => true,
-                '*' => true,
-                '/' => true,
-                '\\' => true,
-                '|' => true,
-                '^' => true,
-                '&' => true,
-                '>' => true,
-                '<' => true,
-                '=' => true,
-                '!' => true,
-                '#' => true,
-                '(' => true,
-                ')' => true,
-                ':' => true,
-                ';' => true,
-                ',' => true,
-                _ => false,
-            };
-        }
-        /*
-         * Calculates <<'s given a specific string
-         */
-        private static string CalculateBitShift(string sINPUT, char chosenType)
-        {
-            if (sINPUT.Contains("<<"))
-            {
-                ulong input;
-                var strings = sINPUT.Split(new string[] { "<<" }, StringSplitOptions.None);
-                if (strings[0] == "")
-                {
-                    strings[0] = "1";
-                }
-                var first = 0;
-                var second = 0ul;
-                int.TryParse(strings[1], out first);
-                ulong.TryParse(strings[0], out second);
-                input = (second << first);
-                PrintColour(string.Format("{0} << {1} = {2}", second, first, input), true);
-                return input.ToString();
-            }
-            else if (sINPUT.Contains(">>"))
-            {
-                ulong input;
-                var strings = sINPUT.Split(new string[] { ">>" }, StringSplitOptions.None);
-                if (strings[0] == "")
-                {
-                    switch (chosenType)
-                    {
-                        case 'b':
-                            strings[0] = (byte.MaxValue / 2 + 1).ToString();
-                            break;
-                        case 's':
-                            strings[0] = (ushort.MaxValue / 2 + 1).ToString();
-                            break;
-                        case 'i':
-                            strings[0] = (uint.MaxValue / 2 + 1).ToString();
-                            break;
-                        default:
-                            strings[0] = (ulong.MaxValue / 2 + 1).ToString();
-                            break;
-                    }
-                }
-                var first = 0;
-                var second = 0ul;
-                int.TryParse(strings[1], out first);
-                ulong.TryParse(strings[0], out second);
-                input = (second >> first);
-                PrintColour(string.Format("{0} >> {1} = {2}", second, first, input), true);
-                return input.ToString();
-            }
-            return "0";
-        }
-        /// <summary>
-        /// Like remove math, except exclusively for bitshifts
-        /// </summary>
-        /// <param name="input"></param>
-        /// <param name="chosenType"></param>
-        /// <returns></returns>
-        public static string RemoveBitShift(string input, char chosenType)
-        {
-            string buffer = "";
-            if (input.Contains("<<") || input.Contains(">>")) //Is there a bitwise operator?
-            {
-                for (int i = 0; i < input.Length; i++) //Iterate through the string
-                {
-                    buffer += input[i];
-                    if (buffer.Contains("<<") || buffer.Contains(">>")) //Did we just get a <<?
-                    {
-                        buffer = ""; //Reset the bugger
-                        int lastOperatorIDX = LastOperatorIDX(input, i - 2);
-                        if ((lastOperatorIDX != 0 || IsOperator(input[0])) && i >= 2)
-                        {
-                            ++lastOperatorIDX;
-                        }
-                        int nextOperatorIDX = NextOperatorIDX(input, i + 1);
-                        string sub = input.Substring(lastOperatorIDX, (nextOperatorIDX - lastOperatorIDX));
-                        string result = CalculateBitShift(sub, chosenType);
-                        string toReturn = string.Format("{0}{1}{2}", input.Substring(0, lastOperatorIDX), result, input.Substring(nextOperatorIDX, input.Length - nextOperatorIDX));
-                        return RemoveBitShift(toReturn, chosenType);
-                    }
-                }
-            }
-            return input;
-        }
         public static void DoLoopFunc(string loop)
         {
             loop = loop.Substring(4);
@@ -2028,270 +878,6 @@ namespace DevTools
             }
         }
 
-        /// <summary>
-        /// VERY BUGGY. WILL BE FIXED LATER
-        /// </summary>
-        /// <param name="input"></param>
-        /// <param name="variableName"></param>
-        /// <param name="variableValue"></param>
-        /// <returns></returns>
-        private static string ReplaceTempVariables(string input, char variableName, string variableValue)
-        {
-            string result = "";
-            for (int i = 0; i < input.Length; ++i)
-            {
-                char c = input[i];
-                if (c == variableName)
-                {
-                    if ((input.Length - 1 == i || IsOperator(input[i + 1])) && (i == 0 || IsOperator(input[i - 1])))
-                    {
-                        result += variableValue;
-                    }
-                    else if (i == 2 && (input.StartsWith("nw") || input.StartsWith("np"))) //Is there a nw or a np command?
-                    {
-                        //Replace the variable
-                        result += variableValue;
-                    }
-                    else if (i == 4 && input.StartsWith("nwnp")) //nw and np?
-                    {
-                        //replace the variable
-                        result += variableValue;
-                    }
-                    else
-                    {
-                        result += c;
-                    }
-                }
-                else
-                {
-                    result += c;
-                }
-            }
-            return result;
-        }
-        /// <summary>
-        /// VERY BUGGY. WILL BE FIXED LATER
-        /// </summary>
-        /// <param name="input"></param>
-        /// <param name="variableName"></param>
-        /// <param name="variableValue"></param>
-        /// <returns></returns>
-        private static string ReplaceTempVariables(string input, string variableName, string variableValue)
-        {
-            string result = "";
-            for (int i = 0; i < input.Length; ++i)
-            {
-                char c = input[i];
-                string currentVar = "";
-                result += c;
-                if (result.Length >= variableName.Length && !IsOperator(input[i]) && !char.IsDigit(input[i]))
-                {
-                    currentVar = result.Substring(i - variableName.Length + 1);
-                }
-                if (currentVar == variableName) //Is this the current variable we are working on
-                {
-                    if (input.Length == result.Length //Are we at the end of the line?
-                    || (IsOperator(input[i + 1]) //At the end of the variable. i.e. there isn't a letter or number after the variable, but an operator
-                    && (i == variableName.Length - 1 || IsOperator(input[i - variableName.Length]))))
-
-                    {
-                        result = result.Substring(0, result.Length - variableName.Length) + input.Substring(i + 1, input.Length - 1 - i);
-                        result = result.Insert(i - variableName.Length + 1, "(" + variableValue + ")");
-                        PrintColour(variableName + " --> " + variableValue, true);
-                        return ReplaceTempVariables(result, variableName, variableValue);
-                    }
-                }
-            }
-            return result;
-        }
-        private static bool ContainsVariable(string input, string variableName)
-        {
-            string result = "";
-            for (int i = 0; i < input.Length; ++i)
-            {
-                char c = input[i];
-                string currentVar = "";
-                result += c;
-                if (result.Length >= variableName.Length)
-                {
-                    currentVar = result.Substring(i - variableName.Length + 1);
-                }
-                if (currentVar == variableName && (input.Length - result.Length == 0 || IsOperator(input[i + 1])) && (i == variableName.Length - 1 || IsOperator(input[i - variableName.Length])))
-                {
-                    return true;
-                }
-            }
-            return false;
-        }
-        /// <summary>
-        /// See if user is modifying a variable that already exists
-        /// </summary>
-        /// <param name="userINPUT">users input</param>
-        /// <returns></returns>
-        public static bool ModifyVariables(string userINPUT)
-        {
-            bool mainresult = false;
-            if (userINPUT.Contains('=') && userINPUT.Substring(0,3) != "var")
-            {
-                Dictionary<string, string> toreplace = new Dictionary<string, string>();
-                List<string> tonull = new List<string>();
-                foreach (var pair in tempVariables)
-                {
-                    if (ContainsVariable(userINPUT, pair.Key))
-                    {
-                        var variableName = pair.Key;
-                        string result = "";
-                        bool add = true;
-                        for (int i = 0; i < userINPUT.Length; ++i)
-                        {
-                            char c = userINPUT[i];
-                            string currentVar = "";
-                            result += c;
-                            if (result.Length >= variableName.Length)
-                            {
-                                currentVar = result.Substring(i - variableName.Length + 1);
-                            }
-                            if (currentVar == variableName && (userINPUT.Length - result.Length == 0 || (userINPUT[i + 1])=='=') && (i == variableName.Length - 1 || (userINPUT[i - variableName.Length])=='='))
-                            {
-                                //The last x characters of result are a variable and the next character is an '=' sign
-                                var value = BitCalculate(RemoveBrackets(ReplaceTempVariables(userINPUT.Substring(i + 2, userINPUT.Length - i - 2)), 'u'), 'u');
-                                if (value != "null")
-                                {
-                                    toreplace.Add(variableName, value);
-                                }
-                                else
-                                {
-                                    tonull.Add(variableName);
-                                }
-                                mainresult = true;
-                                add = false;
-                                break;
-                                //pair.Value = sINPUT.Substring(i + 1, sINPUT.Length - i - 1);
-                            }
-                        }
-                        if (add)
-                        {
-                            toreplace.Add(variableName, tempVariables[variableName]);
-                        }
-                    }
-                }
-                //tempVariables = toreplace;
-                Dictionary<string, string> dresult = new Dictionary<string, string>();
-                foreach (var i in toreplace)    
-                {
-                    dresult.Add(i.Key, i.Value);
-                }
-                foreach (var i in tempVariables)
-                {
-                    if (!dresult.ContainsKey(i.Key) && !tonull.Contains(i.Key))
-                    {
-                        dresult.Add(i.Key, i.Value);
-                    }
-                }
-                tempVariables = dresult;
-            }
-            return mainresult;
-        }
-        public static Dictionary<string, string> tempVariables = new Dictionary<string, string>();
-        public static Dictionary<string, Networking> networkingVariables = new Dictionary<string, Networking>();
-        public static void DefineTempVariable(string variable)
-        {
-            if (variable.Contains("=new")) //Wants to open a port?
-            {
-                string name = variable.Split('=')[0];
-                variable = variable.Substring(name.Length+1+3); //+1 is to remove the equals. +3 is for the 'new'
-
-                string[] variableData = variable.Split('_');
-                ProtocolType protocolType;
-                if (variableData[0] == "tcp") //Starting a TCP server?
-                {
-                    protocolType = ProtocolType.Tcp;
-                }
-                else if(variableData[0] == "udp")
-                {
-                    protocolType = ProtocolType.Udp;
-                }
-                else
-                {
-                    expectingError = true;
-                    throw new Exception("Invalid protocol type: " + variableData[0]);
-                }
-
-                string networkingType = variableData[1].Split('(')[0];
-                if (networkingType == "client") //Creating a client?
-                {
-                    variable = variable.Substring(NextBracket(variable,0)+1); //Remove everything up to the opening bracket
-                    variable = variable.Substring(0,variable.Length-1); //Remove final bracket
-                    string[] data = variable.Split(',');
-                    int port = int.Parse(data[0]);
-                    string address = data[1];
-                    if (address == "localhost")
-                    {
-                        address = "127.0.0.1";
-                    }
-                    try
-                    {
-                        IPAddress.Parse(address);
-                    }
-                    catch
-                    {
-                        expectingError = true;
-                        throw new Exception("Invalid IP address: " + address);
-                    }
-                    ClientNetworking clientNetworking = new ClientNetworking(address, port, NetworkingPrint, protocolType);
-                    networkingVariables.Add(name,clientNetworking);
-                    Thread.Sleep(1000);
-                }
-                else if (networkingType == "server")
-                {
-                    variable = variable.Substring(NextBracket(variable, 0) + 1); //Remove everything up to the opening bracket
-                    variable = variable.Substring(0, variable.Length - 1); //Remove final bracket
-                    int port = int.Parse(variable);
-                    ServerNetworking serverNetworking = new ServerNetworking(port, NetworkingPrint, protocolType);
-                    networkingVariables.Add(name, serverNetworking);
-                }
-                else
-                {
-                    expectingError = true;
-                    throw new Exception("Invalid networking type: " + networkingType);
-                }
-                return;
-            }
-
-            string[] strings = variable.SplitAtFirst('=');
-            int equalsIDX = strings[0].Length - 1;
-            string value = strings[1];
-            string variableName = variable.Substring(0, equalsIDX+1);
-            if (variableName.Any(c => !char.IsLetter(c) && c != ' '))
-            {
-                PrintError("Invalid variable name");
-                return;
-            }
-            if (VariableExists(variableName))
-            {
-                PrintError("Variable is already defined");
-                return;
-            }
-            tempVariables.Add(variableName, value);
-        }
-        public static string ReplaceTempVariables(string sINPUT)
-        {
-            foreach (var pair in networkingVariables)
-            {
-                if (sINPUT.StartsWith(pair.Key)) //Doing operation on a networking thingy?
-                {
-                    string operation = sINPUT.Split('.')[1]; //Find the function after the '.'
-                    DoNetworkingOperation(pair.Value, operation);
-                    return "CLOSE_CONDITION_PROCESSED";
-                }
-            }
-
-            foreach (var pair in tempVariables)
-            {
-                sINPUT = ReplaceTempVariables(sINPUT, pair.Key, pair.Value);
-            }
-            return sINPUT;
-        }
         public static void DoNetworkingOperation(Networking n, string operation)
         {
             if (operation.BeginsWith("send")) //User wants to send data?
@@ -2300,134 +886,6 @@ namespace DevTools
                 data = data.Substring(0,data.Length-1); //Remove send( and )
                 n.Send(data);
             }
-        }
-        private static int NextOperatorIDX(string input, int currIDX)
-        {
-            for (int i = currIDX; i < input.Length; i++)
-            {
-                if (input[i] == '<' ||
-                    input[i] == '>' ||
-                    input[i] == '+' ||
-                    input[i] == '-' ||
-                    input[i] == '*' ||
-                    input[i] == '/' ||
-                    input[i] == '^' ||
-                    input[i] == '&' ||
-                    input[i] == '|' ||
-                    input[i] == '=' ||
-                    char.IsLetter(input[i]))
-                {
-                    return i;
-                }
-            }
-            return input.Length;
-        }
-        private static int NextOperatorIDX_NoLetter(string input, int currIDX)
-        {
-            for (int i = currIDX; i < input.Length; i++)
-            {
-                if (input[i] == '<' ||
-                    input[i] == '>' ||
-                    input[i] == '+' ||
-                    input[i] == '-' ||
-                    input[i] == '*' ||
-                    input[i] == '/' ||
-                    input[i] == '^' ||
-                    input[i] == '&' ||
-                    input[i] == '|' ||
-                    input[i] == ')' ||
-                    input[i] == '(' ||
-                    input[i] == ',')
-                {
-                    return i;
-                }
-            }
-            return input.Length;
-        }
-
-        private static int NextOperatorIDX_NoBrackets(string input, int currIDX)
-        {
-            for (int i = currIDX; i < input.Length; i++)
-            {
-                if (input[i] == '<' ||
-                    input[i] == '>' ||
-                    input[i] == '+' ||
-                    input[i] == '-' ||
-                    input[i] == '*' ||
-                    input[i] == '/' ||
-                    input[i] == '^' ||
-                    input[i] == '&' ||
-                    input[i] == '|' ||
-                    input[i] == ',' ||
-                    char.IsLetter(input[i]))
-                {
-                    return i;
-                }
-            }
-            return input.Length;
-        }
-        private static int NextColonIDX(string input, int currIDX)
-        {
-            for (int i = currIDX; i < input.Length; i++)
-            {
-                if (input[i] == ':')
-                {
-                    return i;
-                }
-            }
-            return input.Length;
-        }
-
-        private static int LastOperatorIDX(string input, int currIDX)
-        {
-            for (int i = currIDX; i > -1; --i)
-            {
-                if (input[i] == '<' ||
-                    input[i] == '>' ||
-                    input[i] == '+' ||
-                    input[i] == '-' ||
-                    input[i] == '*' ||
-                    input[i] == '/' ||
-                    input[i] == '^' ||
-                    input[i] == '&' ||
-                    input[i] == '|' ||
-                    char.IsLetter(input[i]))
-                {
-                    return i;
-                }
-            }
-            return 0;
-        }
-
-        private static int LastAsterixIDX(string input, int currIDX)
-        {
-            for (int i = currIDX; i > -1; --i)
-            {
-                if (input[i] == '*')
-                {
-                    return i;
-                }
-            }
-            return -1;
-        }
-
-        private static int LastNegOperatorIDX(string input, int currIDX)
-        {
-            for (int i = currIDX; i > -1; --i)
-            {
-                if (input[i] == '+' ||
-                    input[i] == '-' ||
-                    input[i] == '*' ||
-                    input[i] == '/' ||
-                    input[i] == '^' ||
-                    input[i] == '&' ||
-                    input[i] == '|' ||
-                    char.IsLetter(input[i]))
-                {
-                    return i;
-                }
-            }
-            return -1;
         }
 
         public static string DoubleToBin(double input, bool flipped = false)
@@ -2565,163 +1023,14 @@ namespace DevTools
         static string DataDirectory = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + @"\DevTools";
 
         static string DataFile = @"\data.txt";
-        static string DataFilePath = DataDirectory + DataFile;
+        public static string DataFilePath = DataDirectory + DataFile;
                       
         static string WorkingsFile = @"\workings.txt";
-        static string WorkingsFilePath = DataDirectory + WorkingsFile;
+        public static string WorkingsFilePath = DataDirectory + WorkingsFile;
 
         static string funcsFile = @"\funcs.txt";
-        static string FuncFilePath = DataDirectory + funcsFile;
-
-        static void DefineVariable(string variable)
-        {
-            string[] strings = variable.SplitAtFirst('=');
-            int equalsIDX = strings[0].Length-1;
-            string value = strings[1];
-            string variableName = variable.Substring(7,equalsIDX-6);
-            if (variableName.Any(c=>!char.IsLetter(c) && c != ' ') )
-            {
-                PrintColour("Invalid variable name", false);
-                return;
-            }
-            if (VariableExists(variableName))
-            {
-                PrintError("Variable is already defined");
-                return;
-            }
-            List<string> contents = DefineVariableContents(variableName, value);
-            if (contents.All(s=>s.SplitAtFirst(',')[0] != variableName))
-            {
-                contents.Add(string.Format("{0},{1}", variableName, value));
-            }
-            File.WriteAllLines(DataFilePath, contents);
-        }
-        public static List<string> DefineVariableContents(string variableName = "", string value = "")
-        {
-            List<string> contents = File.ReadAllLines(DataFilePath).ToList();
-            for (int i = 0; i < contents.Count; i++)
-            {
-                string s = contents[i];
-                string[] args = s.SplitAtFirst(',');
-                if (args[0] == variableName)
-                {
-                    var ss = s.SplitAtFirst(',');
-                    ss[1] = value;
-                    contents[i] = ss[0] + ',' + ss[1];
-                    File.WriteAllLines(DataFilePath, contents);
-                }
-            }
-
-            return contents;
-        }
-        static void DefineFunction(string function)
-        {
-            function = function.Substring("#defunc".Length);
-
-            var prev = File.ReadAllLines(FuncFilePath).ToList();
-            var name = function.Substring(0, function.IndexOf('('));
-            if (VariableExists(name))
-            {
-                PrintError("Variable is already defined");
-                return;
-            }
-            List<string> toremove = new List<string>();
-            foreach (var s in prev)
-            {
-                if (s == "SYSTEM FUNCTIONS:")
-                {
-                    continue;
-                }
-                var bracketidx = s.IndexOf('(');
-                var substr = s.Substring(0,bracketidx);
-                if (substr == name) //Already defined function
-                {
-                    toremove.Add(s);
-                }
-            }
-            foreach (var s in toremove)
-            {
-                prev.Remove(s);
-            }
-            File.WriteAllLines(FuncFilePath, prev.ToArray());
-
-            if (!function.Contains(')'))
-            {
-                expectingError = true;
-                throw new Exception("Wrong formatting for declaring function.\nDid not include closing bracket");
-            }
-            if (function.Split(')')[1].Length == 0)
-            {
-                expectingError = true;
-                throw new Exception("Define an action to take place with the variable");
-            }
-            string result = function + "\n" + File.ReadAllText(FuncFilePath);
-
-            File.WriteAllText(FuncFilePath, result);
-        }
-        static void DeleteFunction(string name)
-        {
-            var prev = File.ReadAllLines(FuncFilePath).ToList();
-            List<string> toremove = new List<string>();
-            foreach (var s in prev)
-            {
-                if (s == "SYSTEM FUNCTIONS:")
-                {
-                    break;
-                }
-                var bracketidx = s.IndexOf('(');
-                var substr = s.Substring(0, bracketidx);
-                if (substr == name) //Already defined function
-                {
-                    toremove.Add(s);
-                }
-            }
-            foreach (var s in toremove)
-            {
-                prev.Remove(s);
-            }
-            File.WriteAllLines(FuncFilePath, prev.ToArray());
-        }
-        static void PrintDescription(string name)
-        {
-            var prev = File.ReadAllLines(FuncFilePath).ToList();
-            var nextidx = 0;
-            for (int i = 0; i < prev.Count; i++)
-            {
-                string? s = prev[i];
-                if (s == "SYSTEM FUNCTIONS:")
-                {
-                    nextidx = i + 1;
-                    break;
-                }
-                var bracketidx = s.IndexOf('(');
-                var substr = s.Substring(0, bracketidx);
-                if (substr == name) //Already defined function
-                {
-                    ShowDescription(s);
-                }
-            }
-            for (int i = nextidx; i < prev.Count; i++)
-            {
-                string? s = prev[i];
-
-                var bracketidx = s.IndexOf('(');
-                var substr = s.Substring(0, bracketidx);
-                if (substr == name) //Already defined function
-                {
-                    ShowDescription(s);
-                }
-            }
-        }
-        static void PrintDescription()
-        {
-            var prev = File.ReadAllLines(FuncFilePath).ToList();
-            foreach (var s in prev)
-            {
-                PrintColour(s + ":");
-                ShowDescription(s);
-            }
-        }
+        public static string FuncFilePath = DataDirectory + funcsFile;
+        
         public static string ReplaceVariables(string input)
         {
 
@@ -2813,358 +1122,6 @@ namespace DevTools
                 result += input[i];
             }
             return result;
-        }
-        public static int NextBracket(string input, int idx)
-        {
-            for (int i = idx + 1; i < input.Length; ++i)
-            {
-                if (input[i] == '(' || input[i] == ')')
-                {
-                    return i;
-                }
-            }
-            return -1;
-        }
-        static string DoubleRemoveBrackets(string s)
-        {
-            string buffer = "";
-            int firstBracketIDX = 0;
-            bool addingToBuffer = false;
-            for (int i = 0; i < s.Length; ++i)
-            {
-                var c = s[i];
-
-                if (c == ')')
-                {
-                    string betweenBrackets = TextBetween(s, firstBracketIDX + 1, i - 1);
-                    string total = DoubleCalculate(betweenBrackets);
-                    string nextString = "";
-                    for (int secondIDX = 0; secondIDX < s.Length; ++secondIDX)
-                    {
-                        if (secondIDX < firstBracketIDX || secondIDX > i)
-                        {
-                            nextString += s[secondIDX];
-                        }
-                        else if (secondIDX == firstBracketIDX)
-                        {
-                            nextString += total;
-                        }
-                    }
-                    if (nextString.Contains('('))
-                    {
-                        return DoubleRemoveBrackets(nextString);
-                    }
-                    else
-                    {
-                        return nextString;
-                    }
-                }
-                else if (addingToBuffer)
-                {
-                    buffer += c;
-                }
-                if (c == '(')
-                {
-                    if (addingToBuffer == true)
-                    {
-                        int nextBracketIDX = NextBracket(s, i);
-                        if (nextBracketIDX == -1)
-                        {
-                            expectingError = true;
-                            throw new Exception("Some weird sh*t is happening");
-                        }
-                        if (s[nextBracketIDX] == ')') //Is this the last layer of brackets?
-                        {
-                            string betweenBrackets = TextBetween(s, i + 1, nextBracketIDX - 1);
-                            string total = DoubleCalculate(betweenBrackets);
-                            string nextString = "";
-                            for (int secondIDX = 0; secondIDX < s.Length; ++secondIDX)
-                            {
-                                if (secondIDX < i || secondIDX > nextBracketIDX)
-                                {
-                                    nextString += s[secondIDX];
-                                }
-                                else if (secondIDX == i)
-                                {
-                                    nextString += total;
-                                }
-                            }
-                            return DoubleRemoveBrackets(nextString);
-                        }
-                        else
-                        {
-                            firstBracketIDX = i;
-                        }
-                    }
-                    else
-                    {
-                        addingToBuffer = true;
-                        firstBracketIDX = i;
-                    }
-                }
-            }
-            return s;
-        }
-        static string RemoveBrackets(string s, char chosenType)
-        {
-            string buffer = "";
-            int firstBracketIDX = 0;
-            bool addingToBuffer = false;
-            for (int i = 0; i < s.Length; ++i)
-            {
-                var c = s[i];
-
-                if (c == ')')
-                {
-                    string betweenBrackets = TextBetween(s, firstBracketIDX + 1, i - 1);
-                    string total = BitCalculate(betweenBrackets, chosenType);
-                    string nextString = "";
-                    for (int secondIDX = 0; secondIDX < s.Length; ++secondIDX)
-                    {
-                        if (secondIDX < firstBracketIDX || secondIDX > i)
-                        {
-                            nextString += s[secondIDX];
-                        }
-                        else if (secondIDX == firstBracketIDX)
-                        {
-                            nextString += total;
-                        }
-                    }
-                    if (nextString.Contains('('))
-                    {
-                        return RemoveBrackets(nextString, chosenType);
-                    }
-                    else
-                    {
-                        return nextString;
-                    }
-                }
-                else if (addingToBuffer)
-                {
-                    buffer += c;
-                }
-                if (c == '(')
-                {
-                    if (addingToBuffer == true)
-                    {
-                        int nextBracketIDX = NextBracket(s, i);
-                        if (nextBracketIDX == -1)
-                        {
-                            expectingError = true;
-                            throw new Exception("Some weird sh*t is happening");
-                        }
-                        if (s[nextBracketIDX] == ')') //Is this the last layer of brackets?
-                        {
-                            string betweenBrackets = TextBetween(s, i + 1, nextBracketIDX - 1);
-                            string total = BitCalculate(betweenBrackets, chosenType);
-                            string nextString = "";
-                            for (int secondIDX = 0; secondIDX < s.Length; ++secondIDX)
-                            {
-                                if (secondIDX < i || secondIDX > nextBracketIDX)
-                                {
-                                    nextString += s[secondIDX];
-                                }
-                                else if (secondIDX == i)
-                                {
-                                    nextString += total;
-                                }
-                            }
-                            return RemoveBrackets(nextString, chosenType);
-                        }
-                        else
-                        {
-                            firstBracketIDX = i;
-                        }
-                    }
-                    else
-                    {
-                        addingToBuffer = true;
-                        firstBracketIDX = i;
-                    }
-                }
-            }
-            return s;
-        }
-        private static string DoubleRemoveMultiplyDivide(string input)
-        {
-            if (input.Contains("*") || input.Contains("/"))
-            {
-                for (int i = 0; i < input.Length; i++)
-                {
-                    if (input[i] =='/' || input[i] == '*')
-                    {
-                        int lastOperatorIDX = LastOperatorIDX(input, i - 2);
-                        int nextOperatorIDX = DoubleNextOperatorIDX(input, i + 1);
-                        string sub = input.Substring(lastOperatorIDX, (nextOperatorIDX - lastOperatorIDX));
-                        string result = DoubleCalculateMultiplyDivide(sub);
-                        string toReturn = string.Format("{0}{1}{2}", input.Substring(0, lastOperatorIDX == 0? 0 : lastOperatorIDX+1), result, input.Substring(nextOperatorIDX, input.Length - nextOperatorIDX));
-                        return DoubleRemoveMultiplyDivide(toReturn);
-                    }
-                }
-            }
-            return input;
-        }
-
-        private static int DoubleNextOperatorIDX(string input, int currIDX)
-        {
-            for (int i = currIDX; i < input.Length; i++)
-            {
-                if ((input[i] == '<' ||
-                    input[i] == '>' ||
-                    input[i] == '+' ||
-                    input[i] == '-' ||
-                    input[i] == '*' ||
-                    input[i] == '/' ||
-                    input[i] == '^' ||
-                    input[i] == '&' ||
-                    input[i] == '|' ||
-                    char.IsLetter(input[i]))
-                    &&(i == 0 | !IsOperator(input[i-1])))
-                {
-                    return i;
-                }
-            }
-            return input.Length;
-        }
-        public static double Average(string sINPUT)
-        {
-            sINPUT = sINPUT.Substring(4);
-            sINPUT = sINPUT.Remove(sINPUT.Length - 1, 1);
-            var snums = sINPUT.Split(',');
-            List<double> doubles = new List<double>();
-            foreach (var s in snums)
-            {
-                doubles.Add(double.Parse(s));
-            }
-            return doubles.Average();
-        }
-        private static string RemoveMultiplyDivide(string input)
-        {
-            if (input.Contains("*") || input.Contains("/"))
-            {
-                for (int i = 0; i < input.Length; i++)
-                {
-                    if (input[i] =='/' || input[i] == '*')
-                    {
-                        int lastOperatorIDX = LastOperatorIDX(input, i - 2);
-                        int nextOperatorIDX = NextOperatorIDX(input, i + 1);
-                        string sub = input.Substring(lastOperatorIDX, (nextOperatorIDX - lastOperatorIDX));
-                        string result = CalculateMultiplyDivide(sub);
-                        string toReturn = string.Format("{0}{1}{2}", input.Substring(0, lastOperatorIDX == 0? 0 : lastOperatorIDX+1), result, input.Substring(nextOperatorIDX, input.Length - nextOperatorIDX));
-                        return RemoveMultiplyDivide(toReturn);
-                    }
-                }
-            }
-            return input;
-        }
-        private static string DoubleCalculateMultiplyDivide(string sINPUT)
-        {
-            if (sINPUT.Contains("*"))
-            {
-                double input;
-                var strings = sINPUT.Split('*');
-                if (strings[0] == "")
-                {
-                    strings[0] = lastInput.ToString();
-                }
-                double first = 0;
-                double second = 0;
-                double.TryParse(strings[1], out first);
-                double.TryParse(strings[0], out second);
-                if (strings[0][0] == '-')
-                {
-                    double.TryParse(strings[0].Substring(1), out second);
-                    second = -second;
-                }
-                if (strings[1][0] == '-')
-                {
-                    double.TryParse(strings[1].Substring(1), out first);
-                    first = -first;
-                }
-                input = second * first;
-                //PrintColour(string.Format("{0} * {1} = {2}", second, first, input), true);
-                return input.ToString();
-            }
-            else if (sINPUT.Contains("/"))
-            {
-                double input;
-                var strings = sINPUT.Split('/');
-                if (strings[0] == "")
-                {
-                    strings[0] = lastInput.ToString();
-                }
-                double first = 0;
-                double second = 0;
-                double.TryParse(strings[1], out first);
-                double.TryParse(strings[0], out second);
-                if (strings[0][0] == '-')
-                {
-                    double.TryParse(strings[0].Substring(1), out second);
-                    second = -second;
-                }
-                if (strings[1][0] == '-')
-                {
-                    double.TryParse(strings[1].Substring(1), out first);
-                    first = -first;
-                }
-                input = second / first;
-                //PrintColour(string.Format("{0} / {1} = {2}", second, first, input), true);
-                return input.ToString();
-            }
-            return "0";
-        }
-        static string prevanswer = "";
-        static int GCD(int[] numbers)
-        {
-            return numbers.Aggregate(GCD);
-        }
-
-        static int GCD(int a, int b)
-        {
-            return b == 0 ? a : GCD(b, a % b);
-        }
-        private static string CalculateMultiplyDivide(string sINPUT)
-        {
-            if (sINPUT.Contains("*"))
-            {
-                ulong input;
-                var strings = sINPUT.Split('*');
-                if (strings[0] == "")
-                {
-                    strings[0] = lastInput.ToString();
-                }
-                var first = 0ul;
-                var second = 0ul;
-                ulong.TryParse(strings[1], out first);
-                ulong.TryParse(strings[0], out second);
-                input = second * first;
-                //PrintColour(string.Format("{0} * {1} = {2}", second, first, input), true);
-                return input.ToString();
-            }
-            else if (sINPUT.Contains("/"))
-            {
-                ulong input;
-                var strings = sINPUT.Split('/');
-                if (strings[0] == "")
-                {
-                    strings[0] = lastInput.ToString();
-                }
-                var first = 0ul;
-                var second = 0ul;
-                ulong.TryParse(strings[1], out first);
-                ulong.TryParse(strings[0], out second);
-                if (first != 0)
-                {
-                    input = second / first;
-                }
-                else
-                {
-                    input = 0;
-                }
-                //PrintColour(string.Format("{0} / {1} = {2}", second, first, input), true);
-                return input.ToString();
-            }
-            return "0";
         }
         /// <summary>
         /// Looks for boolean statements: ==,!=,>,<. Processes their values
@@ -3262,311 +1219,6 @@ namespace DevTools
             }
             result.Add(buffer);
             return result;
-        }
-        private static string RemoveLast(string buffer)
-        {
-            if (buffer == "")
-            {
-                return "";
-            }
-            List<char> chars = buffer.ToArray().ToList();
-            chars.RemoveAt(chars.Count()-1);
-            return new string(chars.ToArray());
-        }
-        private static bool StringContains(string input, char v)
-        {
-            string buffer = "";
-            bool nextCantBeV = false;
-            foreach (var c in input)
-            {
-                if (c==v && nextCantBeV)
-                {
-                    nextCantBeV = false;
-                    buffer += c;
-                    continue;
-                }
-                if (nextCantBeV)
-                {
-                    return true;
-                }
-                if (c == v && buffer.Length != 0 && buffer[buffer.Length-1] != v)
-                {
-                    nextCantBeV = true;
-                }
-                buffer += c;
-            }
-            return false;
-        }
-        private static string BitCalculate(string input, char chosenType)
-        {
-            var sINPUT = RemoveMultiplyDivide(RemoveBitShift(input, chosenType));
-
-            if (sINPUT.Contains('&') || 
-                sINPUT.Contains('|') || 
-                sINPUT.Contains('^') || 
-                sINPUT.Contains('+') || 
-                sINPUT.Contains('-'))
-            {
-                string firstUlong = "";
-                int firstUlongStart_IDX = 0;
-                bool firstTime = true;
-                string secondUlong = "";
-                char _operator = ' ';
-                bool operatorSet = false;
-                for (int i = 0; i < sINPUT.Length; i++)
-                {
-                    char c = sINPUT[i];
-                    if (char.IsNumber(c))
-                    {
-                        //is it a number
-                        if (!operatorSet)
-                        {
-                            firstUlong += c;
-                            if (firstTime)
-                            {
-                                firstTime = false;
-                                firstUlongStart_IDX = i;
-                            }
-                        }
-                        else
-                        {
-                            secondUlong += c;
-                        }
-                    }
-                    else if (c == '|' || c == '&' || c == '^' || c == '+' || c == '-')
-                    {
-                        if (secondUlong != "")
-                        {
-                            ulong result = 0ul;
-                            if (firstUlong == "")
-                            {
-                                firstUlong = lastInput.ToString();
-                                sINPUT = sINPUT.Insert(0, firstUlong);
-                                //firstUlongStart_IDX = 1;
-                            }
-                            switch (_operator)
-                            {
-                                case '&':
-                                    result = ulong.Parse(firstUlong) & ulong.Parse(secondUlong);
-                                    break;
-                                case '|':
-                                    result = ulong.Parse(firstUlong) | ulong.Parse(secondUlong);
-                                    break;
-                                case '^':
-                                    result = ulong.Parse(firstUlong) ^ ulong.Parse(secondUlong);
-                                    break;
-                                case '+':
-                                    result = ulong.Parse(firstUlong) + ulong.Parse(secondUlong);
-                                    break;
-                                case '-':
-                                    result = ulong.Parse(firstUlong) - ulong.Parse(secondUlong);
-                                    break;
-                            }
-                            //PrintColour(sINPUT + " = " + RemoveAndReplace(firstUlongStart_IDX, i, result.ToString(), sINPUT), true);
-                            return BitCalculate(RemoveAndReplace(firstUlongStart_IDX, i, result.ToString(), sINPUT), chosenType);
-                        }
-                        operatorSet = true;
-                        _operator = c;
-                    }
-                    else
-                    {
-                        if (secondUlong != "")
-                        {
-                            ulong result = 0ul;
-                            if (firstUlong == "")
-                            {
-                                firstUlong = lastInput.ToString();
-                                sINPUT = sINPUT.Insert(0, firstUlong);
-                                //firstUlongStart_IDX = 1;
-                            }
-                            switch (_operator)
-                            {
-                                case '&':
-                                    result = ulong.Parse(firstUlong) & ulong.Parse(secondUlong);
-                                    break;
-                                case '|':
-                                    result = ulong.Parse(firstUlong) | ulong.Parse(secondUlong);
-                                    break;
-                                case '^':
-                                    result = ulong.Parse(firstUlong) ^ ulong.Parse(secondUlong);
-                                    break;
-                                case '+':
-                                    result = ulong.Parse(firstUlong) + ulong.Parse(secondUlong);
-                                    break;
-                                case '-':
-                                    result = ulong.Parse(firstUlong) - ulong.Parse(secondUlong);
-                                    break;
-                            }
-                            //PrintColour(sINPUT + " = " + RemoveAndReplace(firstUlongStart_IDX, i, result.ToString(), sINPUT), true);
-                            return BitCalculate(RemoveAndReplace(firstUlongStart_IDX, i, result.ToString(), sINPUT), chosenType);
-                        }
-                        firstUlong = "";
-                        operatorSet = false;
-                        secondUlong = "";
-                    }
-                }
-                if (secondUlong != "")
-                {
-                    ulong result = 0ul;
-                    if (firstUlong == "")
-                    {
-                        firstUlong = lastInput.ToString();
-                        sINPUT = sINPUT.Insert(0, firstUlong);
-                        //firstUlongStart_IDX = 1;
-                    }
-                    ulong first = ulong.Parse(firstUlong);
-                    ulong second = ulong.Parse(secondUlong);
-                    switch (_operator)
-                    {
-                        case '&':
-                            result = ulong.Parse(firstUlong) & ulong.Parse(secondUlong);
-                            break;
-                        case '|':
-                            result = ulong.Parse(firstUlong) | ulong.Parse(secondUlong);
-                            break;
-                        case '^':
-                            result = first ^ second;
-                            break;
-                        case '+':
-                            result = ulong.Parse(firstUlong) + ulong.Parse(secondUlong);
-                            break;
-                        case '-':
-                            result = ulong.Parse(firstUlong) - ulong.Parse(secondUlong);
-                            break;
-                    }
-                    //PrintColour(sINPUT + " = " + RemoveAndReplace(firstUlongStart_IDX, sINPUT.Length, result.ToString(), sINPUT), true);
-                    return BitCalculate(RemoveAndReplace(firstUlongStart_IDX, sINPUT.Length, result.ToString(), sINPUT), chosenType);
-                }
-            }
-            return sINPUT;
-        }
-        private static string DoubleCalculate(string input)
-        {
-            input = Regex.Replace(input, "--", "-");
-            var sINPUT = DoubleRemoveMultiplyDivide(input);
-            if (sINPUT.Where(c=>c=='-').Count() == 1 && !sINPUT.Any(c=>c=='+') && sINPUT[0] == '-')
-            {
-                return sINPUT;
-            }
-            if (sINPUT.Contains('+') ||
-                sINPUT.Contains('-'))
-            {
-                string firstdouble = "";
-                int firstdoubleStart_IDX = 0;
-                bool firstTime = true;
-                string seconddouble = "";
-                char _operator = ' ';
-                bool operatorSet = false;
-                for (int i = 0; i < sINPUT.Length; i++)
-                {
-                    char c = sINPUT[i];
-                    if (char.IsNumber(c) || c == '.')
-                    {
-                        //is it a number
-                        if (!operatorSet)
-                        {
-                            firstdouble += c;
-                            if (firstTime)
-                            {
-                                firstTime = false;
-                                firstdoubleStart_IDX = i;
-                            }
-                        }
-                        else
-                        {
-                            seconddouble += c;
-                        }
-                    }
-                    else if (c == '+' || c == '-')
-                    {
-                        if (seconddouble != "")
-                        {
-                            double result = 0ul;
-                            if (firstdouble == "" && _operator == '-')//negative number
-                            {
-                                operatorSet = true;
-                                firstdouble = '-' + seconddouble;
-                                seconddouble = "";
-                                continue;
-                            }
-                            if (firstdouble == "")
-                            {
-                                firstdouble = lastInput.ToString();
-                                sINPUT = sINPUT.Insert(0, firstdouble);
-                                //firstdoubleStart_IDX = 1;
-                            }
-                            double first = double.Parse(firstdouble);
-                            double second = double.Parse(seconddouble);
-                            switch (_operator)
-                            {
-                                case '+':
-                                    result = first + second;
-                                    break;
-                                case '-':
-                                    result = first - second;
-                                    break;
-                            }
-                            //PrintColour(sINPUT + " = " + RemoveAndReplace(firstdoubleStart_IDX, i, result.ToString(), sINPUT), true);
-                            return DoubleCalculate(RemoveAndReplace(firstdoubleStart_IDX, i, result.ToString(), sINPUT));
-                        }
-                        operatorSet = true;
-                        _operator = c;
-                    }
-                    else
-                    {
-                        if (seconddouble != "")
-                        {
-                            double result = 0ul;
-                            if (firstdouble == "")
-                            {
-                                firstdouble = lastInput.ToString();
-                                sINPUT = sINPUT.Insert(0, firstdouble);
-                                //firstdoubleStart_IDX = 1;
-                            }
-                            double first = double.Parse(firstdouble);
-                            double second = double.Parse(seconddouble);
-                            switch (_operator)
-                            {
-                                case '+':
-                                    result = first + second;
-                                    break;
-                                case '-':
-                                    result = first - second;
-                                    break;
-                            }
-                            //PrintColour(sINPUT + " = " + RemoveAndReplace(firstdoubleStart_IDX, i, result.ToString(), sINPUT), true);
-                            return DoubleCalculate(RemoveAndReplace(firstdoubleStart_IDX, i, result.ToString(), sINPUT));
-                        }
-                        firstdouble = "";
-                        operatorSet = false;
-                        seconddouble = "";
-                    }
-                }
-                if (seconddouble != "")
-                {
-                    double result = 0ul;
-                    if (firstdouble == "")
-                    {
-                        firstdouble = lastInput.ToString();
-                        sINPUT = sINPUT.Insert(0, firstdouble);
-                        //firstdoubleStart_IDX = 1;
-                    }
-                    double first = double.Parse(firstdouble);
-                    double second = double.Parse(seconddouble);
-                    switch (_operator)
-                    {
-                        case '+':
-                            result = first + second;
-                            break;
-                        case '-':
-                            result = first - second;
-                            break;
-                    }
-                    //PrintColour(sINPUT + " = " + RemoveAndReplace(firstdoubleStart_IDX, sINPUT.Length, result.ToString(), sINPUT), true);
-                    return DoubleCalculate(RemoveAndReplace(firstdoubleStart_IDX, sINPUT.Length, result.ToString(), sINPUT));
-                }
-            }
-            return sINPUT;
         }
 
         public static string UlongToBin(ulong input, bool flipped)
